# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 知序云 API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /student/workspace:
    get:
      operationId: getStudentWorkspace
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 当前学生完整工作台数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentWorkspace' }
        '403': { description: 需要学生角色 }
  /student/tasks/{taskId}/text-submission:
    post:
      operationId: submitStudentText
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1 }
      responses:
        '200':
          description: 提交后返回最新学生工作台
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentWorkspace' }
  /student/tasks/{taskId}/file-submission:
    post:
      operationId: submitStudentFile
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: 提交后返回最新学生工作台
          content:
            application/json:
                schema: { $ref: '#/components/schemas/StudentWorkspace' }
  /student/tasks/{taskId}/answer-submission:
    post:
      operationId: submitStudentAnswers
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers: { type: object, additionalProperties: true }
      responses:
        '200': { description: 结构化作业答案已提交并返回最新学生工作台 }
        '400': { description: 题目未完成或答题格式错误 }
  /student/resources/{resourceId}/download:
    get:
      operationId: downloadStudentCourseResource
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: resourceId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: 下载当前学生所属课程的资料 }
        '404': { description: 资料不存在或不属于当前学生课程 }
  /student/assistant/ask:
    post:
      operationId: askStudentAssistant
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1 }
      responses:
        '200':
          description: 学生助手回复
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
        '503': { description: 未配置模型 API Key }
  /student/assistant/sessions:
    post:
      operationId: createStudentAssistantSession
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 已创建当前学生私有 AI 对话 }
  /student/assistant/sessions/{sessionId}/messages:
    post:
      operationId: sendStudentAssistantMessage
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties: { content: { type: string, minLength: 1 } }
      responses:
        '200': { description: 已保存当前学生消息与 AI 回复 }
        '404': { description: 对话不存在或不属于当前学生 }
  /student/teacher-contacts:
    get:
      operationId: listStudentTeacherContacts
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 当前学生课程的任课教师列表 }
  /student/conversations:
    post:
      operationId: createStudentConversation
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teacherId, content]
              properties:
                teacherId: { type: integer, format: int64 }
                content: { type: string, minLength: 1 }
      responses:
        '200': { description: 已创建或继续与任课教师的会话 }
        '404': { description: 教师不属于当前学生课程 }
  /student/conversations/{conversationId}/messages:
    post:
      operationId: sendStudentConversationMessage
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: conversationId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties: { content: { type: string, minLength: 1 } }
      responses:
        '200': { description: 学生消息已发送 }
        '404': { description: 会话不存在或不属于当前学生 }
  /student/conversations/{conversationId}/read:
    post:
      operationId: readStudentConversation
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: conversationId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: 当前学生会话已读 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    StudentWorkspace:
      type: object
      required: [profile, metrics, courses, tasks, reports, notifications, assistantPrompts, assistantSessions, conversations, teacherContacts]
      properties:
        profile: { type: object, additionalProperties: true }
        metrics: { type: object, additionalProperties: true }
        courses: { type: array, items: { type: object, additionalProperties: true } }
        tasks: { type: array, items: { type: object, additionalProperties: true } }
        reports: { type: array, items: { type: object, additionalProperties: true } }
        notifications: { type: array, items: { type: object, additionalProperties: true } }
        assistantPrompts: { type: array, items: { type: string } }
        assistantSessions: { type: array, items: { type: object, additionalProperties: true } }
        conversations: { type: array, items: { type: object, additionalProperties: true } }
        teacherContacts: { type: array, items: { type: object, additionalProperties: true } }
...